requirements.txt
asgiref==3.11.0
certifi==2026.1.4
distlib==0.4.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
packaging==26.0
pillow==12.1.0
pipenv==2026.0.3
platformdirs==4.5.1
PyJWT==2.11.0
setuptools==82.0.0
sqlparse==0.5.5
tzdata==2025.3
virtualenv==20.36.1
